+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-04T15:30:52.121Z",
+    "2026-05-05T08:17:15.795Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {