+++ packages/orq-rc/src/models/operations/retrievedatasource.ts
@@ -118,7 +118,7 @@ export const RetrieveDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KQSSTN55GDVW1V74DQ93P6ST"),
+  _id: z.string().default("01KQVKDCKSKTVB45F5VE2CYJW3"),
   display_name: z.string(),
   description: z.string().optional(),
   status: RetrieveDatasourceStatus$inboundSchema,