+++ packages/orq-rc/RELEASES.md
@@ -8849,3 +8849,13 @@ Based on:
 - [typescript v4.9.0-rc.22] packages/orq-rc
 ### Releases
 - [NPM v4.9.0-rc.22] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.22 - packages/orq-rc
+
+## 2026-05-05 08:20:43
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.11 (2.881.17) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.9.0-rc.24] packages/orq-rc
+### Releases
+- [NPM v4.9.0-rc.24] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.24 - packages/orq-rc